Expose the extracted kv-pair to the outside world so people can look at the cbor bytes
